Return logics with updated client, move getRating to getters
