+++ package.json
@@ -17,7 +17,7 @@
     "clsx": "^2.1.1",
     "radix-vue": "^1.9.10",
     "tailwind-merge": "^2.5.5",
-    "tailwindcss-animate": "^1.0.7",
+    "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
     "@pinia/testing": "^0.1.7",
